+++ resources/js/views/Shared.vue
@@ -64,6 +64,7 @@ import { mapGetters } from 'vuex'
 import { events } from '../bus'
 import DesktopSharepageToolbar from '../components/Layout/Toolbars/DesktopSharepageToolbar'
 import RemoteUploadPopup from "../components/RemoteUpload/RemoteUploadPopup";
+import {getFilesFromDataTransferItems} from "datatransfer-files-promise";
 
 export default {
     name: 'Shared',
@@ -95,8 +96,14 @@ export default {
         }
     },
     methods: {
-		uploadDroppedItems(event) {
-			this.$uploadDraggedFiles(event, this.currentFolder?.data.id)
+		async uploadDroppedItems(event) {
+			// Check if user dropped folder with files
+			let files = await getFilesFromDataTransferItems(event.dataTransfer.items)
+
+			if (files.length !== 0) {
+				// Upload folder with files
+				this.$uploadDraggedFolderOrFile(files, this.currentFolder?.data.id)
+			}
 		},
         contextMenu(event, item) {
             events.$emit('context-menu:show', event, item)