+++ resources/js/views/UploadRequest.vue
@@ -56,6 +56,7 @@ import DragUI from '../components/UI/Others/DragUI'
 import { mapGetters } from 'vuex'
 import { events } from '../bus'
 import RemoteUploadPopup from "../components/RemoteUpload/RemoteUploadPopup";
+import {getFilesFromDataTransferItems} from "datatransfer-files-promise";
 
 export default {
     name: 'UploadRequest',
@@ -87,8 +88,14 @@ export default {
         }
     },
     methods: {
-		uploadDroppedItems(event) {
-			this.$uploadDraggedFiles(event, this.currentFolder?.data.id)
+		async uploadDroppedItems(event) {
+			// Check if user dropped folder with files
+			let files = await getFilesFromDataTransferItems(event.dataTransfer.items)
+
+			if (files.length !== 0) {
+				// Upload folder with files
+				this.$uploadDraggedFolderOrFile(files, this.currentFolder?.data.id)
+			}
 		},
         contextMenu(event, item) {
             events.$emit('context-menu:show', event, item)