+++ changelog.md
@@ -4,6 +4,8 @@
 - Fixed issue when after deleting user, the related subscription wasn't deleted
 - Fixed issue when you perform composer update with private repository
 - Fixed issue where change in sorting option will duplicate the content in file view
+- Fixed issue where Dragged & Dropped folder from desktop didn't start uploading
+
 ## Version 2.2.0.6
 #### Release date: 13. Jun 2022
 - Solved issue when user wasn't redirected into stripe checkout